requirements.txt
Flask>=2.0
openai>=1.0.0
python-dotenv>=0.21
